+++ cash/src/test/java/ru/yandex/practicum/cash/service/CashServiceIntegrationTest.java
@@ -8,6 +8,8 @@ import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import ru.yandex.practicum.cash.dto.CashOperationRequest;
 import ru.yandex.practicum.cash.dto.CashOperationResponse;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.*;
@@ -34,10 +36,10 @@ class CashServiceIntegrationTest {
 
     @Test
     void cashOperationDepositSuccess() {
-        CashOperationRequest request = new CashOperationRequest(1L, "DEPOSIT", "RUB", 1000.0);
+        CashOperationRequest request = new CashOperationRequest(1L, "DEPOSIT", "RUB", BigDecimal.valueOf(1000.0));
 
         when(blockerClient.checkOperation(any(CashOperationRequest.class))).thenReturn(false);
-        when(accountsClient.deposit(1L, "RUB", 1000.0)).thenReturn(5000.0);
+        when(accountsClient.deposit(1L, "RUB", BigDecimal.valueOf(1000.0))).thenReturn(BigDecimal.valueOf(5000.0));
 
         CashOperationResponse response = cashService.processCashOperation(request);
 
@@ -46,16 +48,16 @@ class CashServiceIntegrationTest {
         assertEquals(5000.0, response.newBalance());
 
         verify(blockerClient).checkOperation(any(CashOperationRequest.class));
-        verify(accountsClient).deposit(1L, "RUB", 1000.0);
-        verify(notificationsClient).sendNotification(eq(1L), eq("DEPOSIT"), anyString(), eq(1000.0), eq("RUB"));
+        verify(accountsClient).deposit(1L, "RUB", BigDecimal.valueOf(1000.0));
+        verify(notificationsClient).sendNotification(eq(1L), eq("DEPOSIT"), anyString(), BigDecimal.valueOf(eq(1000.0)), eq("RUB"));
     }
 
     @Test
     void cashOperationWithdrawSuccess() {
-        CashOperationRequest request = new CashOperationRequest(1L, "WITHDRAW", "RUB", 500.0);
+        CashOperationRequest request = new CashOperationRequest(1L, "WITHDRAW", "RUB", BigDecimal.valueOf(500));
 
         when(blockerClient.checkOperation(any(CashOperationRequest.class))).thenReturn(false);
-        when(accountsClient.withdraw(1L, "RUB", 500.0)).thenReturn(3500.0);
+        when(accountsClient.withdraw(1L, "RUB", BigDecimal.valueOf(500.0))).thenReturn(BigDecimal.valueOf(3500.0));
 
         CashOperationResponse response = cashService.processCashOperation(request);
 
@@ -64,13 +66,13 @@ class CashServiceIntegrationTest {
         assertEquals(3500.0, response.newBalance());
 
         verify(blockerClient).checkOperation(any(CashOperationRequest.class));
-        verify(accountsClient).withdraw(1L, "RUB", 500.0);
-        verify(notificationsClient).sendNotification(eq(1L), eq("WITHDRAW"), anyString(), eq(500.0), eq("RUB"));
+        verify(accountsClient).withdraw(1L, "RUB", BigDecimal.valueOf(500.0));
+        verify(notificationsClient).sendNotification(eq(1L), eq("WITHDRAW"), anyString(), BigDecimal.valueOf(eq(500.0)), eq("RUB"));
     }
 
     @Test
     void cashOperationWhenBlocked() {
-        CashOperationRequest request = new CashOperationRequest(1L, "WITHDRAW", "RUB", 150000.0);
+        CashOperationRequest request = new CashOperationRequest(1L, "WITHDRAW", "RUB", BigDecimal.valueOf(150000.0));
 
         when(blockerClient.checkOperation(any(CashOperationRequest.class))).thenReturn(true);
 
@@ -81,17 +83,17 @@ class CashServiceIntegrationTest {
         assertNull(response.newBalance());
 
         verify(blockerClient).checkOperation(any(CashOperationRequest.class));
-        verify(accountsClient, never()).deposit(anyLong(), anyString(), anyDouble());
-        verify(accountsClient, never()).withdraw(anyLong(), anyString(), anyDouble());
-        verify(notificationsClient, never()).sendNotification(anyLong(), anyString(), anyString(), anyDouble(), anyString());
+        verify(accountsClient, never()).deposit(anyLong(), anyString(), BigDecimal.valueOf(anyDouble()));
+        verify(accountsClient, never()).withdraw(anyLong(), anyString(), BigDecimal.valueOf(anyDouble()));
+        verify(notificationsClient, never()).sendNotification(anyLong(), anyString(), anyString(), BigDecimal.valueOf(anyDouble()), anyString());
     }
 
     @Test
     void cashOperationWhenAccountsClientFails() {
-        CashOperationRequest request = new CashOperationRequest(1L, "DEPOSIT", "RUB", 1000.0);
+        CashOperationRequest request = new CashOperationRequest(1L, "DEPOSIT", "RUB", BigDecimal.valueOf(1000.0));
 
         when(blockerClient.checkOperation(any(CashOperationRequest.class))).thenReturn(false);
-        when(accountsClient.deposit(1L, "RUB", 1000.0))
+        when(accountsClient.deposit(1L, "RUB", BigDecimal.valueOf(1000.0)))
                 .thenThrow(new RuntimeException("Insufficient funds"));
 
         CashOperationResponse response = cashService.processCashOperation(request);
@@ -101,18 +103,18 @@ class CashServiceIntegrationTest {
         assertNull(response.newBalance());
 
         verify(blockerClient).checkOperation(any(CashOperationRequest.class));
-        verify(accountsClient).deposit(1L, "RUB", 1000.0);
-        verify(notificationsClient, never()).sendNotification(anyLong(), anyString(), anyString(), anyDouble(), anyString());
+        verify(accountsClient).deposit(1L, "RUB", BigDecimal.valueOf(1000.0));
+        verify(notificationsClient, never()).sendNotification(anyLong(), anyString(), anyString(), BigDecimal.valueOf(anyDouble()), anyString());
     }
 
     @Test
     void cashOperationWhenNotificationFails() {
-        CashOperationRequest request = new CashOperationRequest(1L, "DEPOSIT", "RUB", 1000.0);
+        CashOperationRequest request = new CashOperationRequest(1L, "DEPOSIT", "RUB", BigDecimal.valueOf(1000.0));
 
         when(blockerClient.checkOperation(any(CashOperationRequest.class))).thenReturn(false);
-        when(accountsClient.deposit(1L, "RUB", 1000.0)).thenReturn(5000.0);
+        when(accountsClient.deposit(1L, "RUB", BigDecimal.valueOf(1000.0))).thenReturn(BigDecimal.valueOf(5000.0));
         doThrow(new RuntimeException("Notification service unavailable"))
-                .when(notificationsClient).sendNotification(anyLong(), anyString(), anyString(), anyDouble(), anyString());
+                .when(notificationsClient).sendNotification(anyLong(), anyString(), anyString(), BigDecimal.valueOf(anyDouble()), anyString());
 
         CashOperationResponse response = cashService.processCashOperation(request);
 
@@ -121,12 +123,12 @@ class CashServiceIntegrationTest {
         assertEquals(5000.0, response.newBalance());
 
         verify(blockerClient).checkOperation(any(CashOperationRequest.class));
-        verify(accountsClient).deposit(1L, "RUB", 1000.0);
+        verify(accountsClient).deposit(1L, "RUB", BigDecimal.valueOf(1000.0));
     }
 
     @Test
     void cashOperationWithInvalidOperationType() {
-        CashOperationRequest request = new CashOperationRequest(1L, "INVALID", "RUB", 1000.0);
+        CashOperationRequest request = new CashOperationRequest(1L, "INVALID", "RUB", BigDecimal.valueOf(1000.0));
 
         when(blockerClient.checkOperation(any(CashOperationRequest.class))).thenReturn(false);
 
@@ -137,20 +139,20 @@ class CashServiceIntegrationTest {
         assertNull(response.newBalance());
 
         verify(blockerClient).checkOperation(any(CashOperationRequest.class));
-        verify(accountsClient, never()).deposit(anyLong(), anyString(), anyDouble());
-        verify(accountsClient, never()).withdraw(anyLong(), anyString(), anyDouble());
+        verify(accountsClient, never()).deposit(anyLong(), anyString(), BigDecimal.valueOf(anyDouble()));
+        verify(accountsClient, never()).withdraw(anyLong(), anyString(), BigDecimal.valueOf(anyDouble()));
     }
 
     @Test
     void cashOperationWhenBlockerServiceFails() {
-        CashOperationRequest request = new CashOperationRequest(1L, "DEPOSIT", "RUB", 1000.0);
+        CashOperationRequest request = new CashOperationRequest(1L, "DEPOSIT", "RUB", BigDecimal.valueOf(1000.0));
 
         when(blockerClient.checkOperation(any(CashOperationRequest.class))).thenReturn(false);
-        when(accountsClient.deposit(1L, "RUB", 1000.0)).thenReturn(5000.0);
+        when(accountsClient.deposit(1L, "RUB", BigDecimal.valueOf(1000.0))).thenReturn(BigDecimal.valueOf(5000.0));
 
         CashOperationResponse response = cashService.processCashOperation(request);
 
         assertTrue(response.success());
-        verify(accountsClient).deposit(1L, "RUB", 1000.0);
+        verify(accountsClient).deposit(1L, "RUB", BigDecimal.valueOf(1000.0));
     }
 }
